requirements.txt
# Core
numpy
requests
paho-mqtt

# Face recognition
opencv-python-headless
# Use opencv-python if you need a display locally

# UI (on device only)
PyQt5

# LAN receiver server (on PC only)
fastapi
uvicorn
python-multipart
pydantic

# sqlite3 is built into Python — no extra install needed
# mysql-connector-python removed (no longer required)
